replaced int64_t

+++ Math/main.c
@@ -8,7 +8,7 @@ int main(int argc, char* argv[])
 
     push_back(&vekt, (void *) 2147483648);
     for (size_t i = 0; i < vekt.size; ++i)
-        printf("%li\n", (int64_t) vekt.value[i]);
+        printf("%li\n", (long long) vekt.value[i]);
 
     return 0;
 }
